move osn2_sim_state and distorted_fcc_al_conventional_sim_state fixtures to root conftest
